Upgrade azure-blob-storage dependencies
Upgrade to the newest stable versions of the azure storage SDK. The
versions numbers of the batch client and the standard client have
diverged, but since azure-storage-blob-batch:12.11.0
transitively depends on azure-storage-blob:12.14.0, the dependencies are
aligned.

+++ ambry-cloud/src/main/java/com/github/ambry/cloud/azure/StorageClient.java
@@ -344,7 +344,7 @@ public abstract class StorageClient {
     HttpClient client = new NettyAsyncHttpClientBuilder().proxy(proxyOptions).build();
 
     // Note: retry decisions are made at CloudBlobStore level.  Configure storageClient with no retries.
-    RequestRetryOptions noRetries = new RequestRetryOptions(RetryPolicyType.FIXED, 1, null, null, null, null);
+    RequestRetryOptions noRetries = new RequestRetryOptions(RetryPolicyType.FIXED, 1, (Integer) null, null, null, null);
     try {
       return buildBlobServiceClient(client, storageConfiguration, noRetries, azureCloudConfig);
     } catch (MalformedURLException | InterruptedException | ExecutionException ex) {
